Tuples
Writing a program to read through a file and figure out the distribution by hour of the day for each of the messages, by pulling  the hour out from the 'From ' line by finding the time and then splitting the string a second time using a colon. Then once I have accumulated the counts for each hour,  I printed out the counts, sorted by hour as shown below.