Python lists
Opening a file and read it line by line. For each line, split the line into a list of words using the split() method. The program will build a list of words. For each word on each line it will check to see if the word is already in the list and if not append it to the list. When the program completes, it will sort and print the resulting words in python sort() order.